Add DA3 giant model reference with pose, rays, and gaussians verification
Extend the comparison framework to support the DA3-Giant model (1.2B params,
ViT-G with SwiGLU) and verify all output modalities against our CUDA
implementation.

Results (bf16 PyTorch vs fp16 CUDA through 40-block backbone):
- Depth:     Pearson r=0.949
- Pose:      Quaternion cosine similarity=0.971
- Gaussians: Per-channel Pearson r=0.997-0.999, overall r=0.998
- Rays:      GPU produces zeros (needs camera conditioning), ref non-zero

New files:
- ref/da3/run_reference_giant.py: Giant model with CameraDec, AuxDPT, GSDPT

Modified files:
- cuda/da3/test_cuda_da3.c: Add --npy-dir flag for saving pose/rays/gaussians
  as .npy, add write_npy_f32_3d() and write_npy_f32_1d() helpers
- ref/da3/run_reference.py: Fix softmax precision for bf16 compatibility

+++ ref/da3/run_reference.py
@@ -88,7 +88,7 @@ class Attention(nn.Module):
             k = torch.cat([k_cls, k_sp], dim=2)
 
         attn = torch.matmul(q * self.scale, k.transpose(-2, -1))
-        attn = F.softmax(attn, dim=-1)
+        attn = F.softmax(attn, dim=-1, dtype=torch.float32).to(v.dtype)
         x = torch.matmul(attn, v)
         x = x.transpose(1, 2).reshape(B, N, C)
         x = self.proj(x)
